Share directory tool launch log panel
