+++ Sources/App/TerminalDirectoryOpenSupport.swift
@@ -490,7 +490,11 @@ final class VSCodeServeWebController {
     }
 #endif
 
-    func ensureServeWebURL(vscodeApplicationURL: URL, completion: @escaping (URL?) -> Void) {
+    func ensureServeWebURL(
+        vscodeApplicationURL: URL,
+        progress: ((String) -> Void)? = nil,
+        completion: @escaping (URL?) -> Void
+    ) {
         queue.async {
             if let process = self.serveWebProcess,
                process.isRunning,
@@ -523,7 +527,8 @@ final class VSCodeServeWebController {
                 }
                 let launchResult = self.launchServeWebProcess(
                     vscodeApplicationURL: vscodeApplicationURL,
-                    expectedGeneration: launchGeneration
+                    expectedGeneration: launchGeneration,
+                    progress: progress
                 )
                 self.queue.async {
                     guard self.activeLaunchGeneration == launchGeneration else {
@@ -626,7 +631,8 @@ final class VSCodeServeWebController {
 
     private func launchServeWebProcess(
         vscodeApplicationURL: URL,
-        expectedGeneration: UInt64
+        expectedGeneration: UInt64,
+        progress: ((String) -> Void)?
     ) -> (process: Process, url: URL)? {
         if let launchProcessOverride {
             return launchProcessOverride(vscodeApplicationURL, expectedGeneration)
@@ -656,7 +662,12 @@ final class VSCodeServeWebController {
         process.standardOutput = stdoutPipe
         process.standardError = stderrPipe
 
-        let collector = ServeWebOutputCollector()
+        let collector = ServeWebOutputCollector { output in
+            guard let progress else { return }
+            DispatchQueue.main.async {
+                progress(output)
+            }
+        }
         let outputReader: (FileHandle) -> Void = { fileHandle in
             switch ProcessPipeReader.readAvailableDataOrEndOfFile(from: fileHandle) {
             case .data(let data):
@@ -794,11 +805,18 @@ final class VSCodeServeWebController {
 }
 
 final class ServeWebOutputCollector {
+    private static let maxOutputSnippetLength = 800
+
     private let lock = NSLock()
     private let semaphore = DispatchSemaphore(value: 0)
+    private let outputHandler: ((String) -> Void)?
     private var outputBuffer = ""
     private var resolvedURL: URL?
     private var didSignal = false
+
+    init(outputHandler: ((String) -> Void)? = nil) {
+        self.outputHandler = outputHandler
+    }
 
     var webUIURL: URL? {
         lock.lock()
@@ -806,39 +824,76 @@ final class ServeWebOutputCollector {
         return resolvedURL
     }
 
-    func append(_ data: Data) {
-        guard let text = String(data: data, encoding: .utf8), !text.isEmpty else { return }
+    var outputSnippet: String {
         lock.lock()
         defer { lock.unlock() }
-        guard resolvedURL == nil else { return }
-        outputBuffer.append(text)
-        while let newlineIndex = outputBuffer.firstIndex(where: \.isNewline) {
-            let line = String(outputBuffer[..<newlineIndex])
-            outputBuffer.removeSubrange(...newlineIndex)
-            guard let parsedURL = VSCodeServeWebURLBuilder.extractWebUIURL(from: line) else {
-                continue
+        return Self.snippet(from: outputBuffer)
+    }
+
+    private static func snippet(from output: String) -> String {
+        let trimmed = output.trimmingCharacters(in: .whitespacesAndNewlines)
+        guard trimmed.count > maxOutputSnippetLength else { return trimmed }
+        let suffix = trimmed.suffix(maxOutputSnippetLength)
+        return "..." + suffix
+    }
+
+    func append(_ data: Data) {
+        guard let text = String(data: data, encoding: .utf8), !text.isEmpty else { return }
+        var outputToReport: String?
+        var shouldSignal = false
+
+        lock.lock()
+        if resolvedURL == nil {
+            outputBuffer.append(text)
+            outputToReport = Self.snippet(from: outputBuffer)
+            while let newlineIndex = outputBuffer.firstIndex(where: \.isNewline) {
+                let line = String(outputBuffer[..<newlineIndex])
+                outputBuffer.removeSubrange(...newlineIndex)
+                guard let parsedURL = VSCodeServeWebURLBuilder.extractWebUIURL(from: line) else {
+                    continue
+                }
+                resolvedURL = parsedURL
+                outputBuffer.removeAll(keepingCapacity: false)
+                if !didSignal {
+                    didSignal = true
+                    shouldSignal = true
+                }
+                break
             }
-            resolvedURL = parsedURL
-            outputBuffer.removeAll(keepingCapacity: false)
-            if !didSignal {
-                didSignal = true
-                semaphore.signal()
-            }
-            return
+        }
+        lock.unlock()
+
+        if let outputToReport {
+            outputHandler?(outputToReport)
+        }
+        if shouldSignal {
+            semaphore.signal()
         }
     }
 
     func markProcessExited() {
+        var outputToReport: String?
+        var shouldSignal = false
+
         lock.lock()
-        defer { lock.unlock() }
         if resolvedURL == nil, !outputBuffer.isEmpty,
            let parsedURL = VSCodeServeWebURLBuilder.extractWebUIURL(from: outputBuffer) {
             resolvedURL = parsedURL
+            outputToReport = Self.snippet(from: outputBuffer)
             outputBuffer.removeAll(keepingCapacity: false)
         }
-        guard !didSignal else { return }
-        didSignal = true
-        semaphore.signal()
+        if !didSignal {
+            didSignal = true
+            shouldSignal = true
+        }
+        lock.unlock()
+
+        if let outputToReport {
+            outputHandler?(outputToReport)
+        }
+        if shouldSignal {
+            semaphore.signal()
+        }
     }
 
     func waitForURL(timeoutSeconds: TimeInterval) -> Bool {
@@ -1189,8 +1244,8 @@ final class DirectoryToolWebServerOutputCollector {
         if resolvedURL == nil,
            let parsedURL = DirectoryToolWebServerURLBuilder.extractURL(from: outputBuffer, pattern: urlPattern) {
             resolvedURL = parsedURL
-            outputBuffer.removeAll(keepingCapacity: false)
             outputToReport = Self.snippet(from: outputBuffer)
+            outputBuffer.removeAll(keepingCapacity: false)
         }
         if !didSignal {
             didSignal = true
