+++ Sources/AppDelegate.swift
@@ -7281,15 +7281,54 @@ final class AppDelegate: NSObject, NSApplicationDelegate, UNUserNotificationCent
             ?? targetTabManager.addWorkspace(select: true).id
         let normalizedDirectoryURL = directoryURL.standardizedFileURL
 
-        VSCodeServeWebController.shared.ensureServeWebURL(vscodeApplicationURL: vscodeApplicationURL) { serveWebURL in
+        var progressController: DirectoryToolLaunchPanelController!
+        progressController = DirectoryToolLaunchPanelController(
+            title: DirectoryToolLaunchPanelController.startingTitle(displayName: "VS Code"),
+            initialMessage: String(
+                localized: "directoryTool.launchProgress.message",
+                defaultValue: "Waiting for the tool to print a local URL."
+            ),
+            launchingMessage: String(
+                localized: "directoryTool.launchProgress.message",
+                defaultValue: "Waiting for the tool to print a local URL."
+            ),
+            initialOutput: String(
+                localized: "directoryTool.launchProgress.noOutput",
+                defaultValue: "No output yet."
+            ),
+            requiresApproval: false,
+            onAllow: nil,
+            onStop: {
+                VSCodeServeWebController.shared.stop()
+            }
+        )
+        progressController.show(presentingWindow: NSApp.keyWindow ?? NSApp.mainWindow)
+
+        VSCodeServeWebController.shared.ensureServeWebURL(
+            vscodeApplicationURL: vscodeApplicationURL,
+            progress: { output in
+                progressController.updateOutput(output)
+            }
+        ) { serveWebURL in
             guard let serveWebURL,
                   let openFolderURL = VSCodeServeWebURLBuilder.openFolderURL(
                       baseWebUIURL: serveWebURL,
                       directoryPath: normalizedDirectoryURL.path
                   ) else {
+                progressController.finish(
+                    message: String(
+                        localized: "directoryTool.launchProgress.failed",
+                        defaultValue: "The server did not start."
+                    ),
+                    stopTitle: String(
+                        localized: "directoryTool.launchProgress.hide",
+                        defaultValue: "Hide"
+                    )
+                )
                 NSSound.beep()
                 return
             }
+            progressController.close()
 
             guard targetTabManager.openBrowser(
                 inWorkspace: targetWorkspaceId,
